# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ECHO Mind Path A API",
    "description": "心理健康记录、筛查提示、审核练习、人工接管与审计。非诊断、非紧急服务替代。",
    "version": "0.6.0"
  },
  "paths": {
...
    "/v1/skills": {
      "get": {
        "summary": "List Skills",
        "description": "用户拉取已 signed 的 Skill 列表（脱敏后下发；draft/reviewed/retired 不下发）。",
        "operationId": "list_skills_v1_skills_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/skills/{skill_id}": {
      "get": {
        "summary": "Get Skill",
        "description": "用户拉取单个 Skill 详情（脱敏后下发；signed-only + action_type 白名单）。",
        "operationId": "get_skill_v1_skills__skill_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Skill Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/skills/completions": {
      "post": {
        "summary": "Create Skill Completion",
        "description": "Skill 执行完成/停止上报（幂等：同 tenant+event_id 返回同一记录）。\n\n- 校验 skill 属于当前用户且 status=='signed'\n- audit action=skill.completion\n- 服务端只记录执行状态，不承载可执行内容",
        "operationId": "create_skill_completion_v1_skills_completions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillCompletionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/skills/{skill_id}/transition": {
      "post": {
        "summary": "Transition Skill",
        "description": "Skill 治理状态机转换：draft→reviewed→signed→retired（不能跳级、不能逆转）。\n\n转入 signed 时（PRD 契约点 5）必须提供 policy_version 与 review_evidence（缺失 → 422）；\ncontent_hash 缺失视为无效内容（不可签署）。",
        "operationId": "transition_skill_v1_skills__skill_id__transition_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Skill Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillTransition"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SkillCompletionCreate": {
        "properties": {
          "event_id": {
            "type": "string",
            "maxLength": 80,
            "minLength": 8,
            "title": "Event Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "skill_id": {
            "type": "string",
            "title": "Skill Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "started",
              "completed",
              "stopped"
            ],
            "title": "Status"
          },
          "duration_seconds": {
            "type": "integer",
            "maximum": 86400.0,
            "minimum": 0.0,
            "title": "Duration Seconds",
            "default": 0
          },
          "client_time": {
            "type": "string",
            "format": "date-time",
            "title": "Client Time"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "user_id",
          "skill_id",
          "status",
          "client_time"
        ],
        "title": "SkillCompletionCreate",
        "description": "Skill 执行完成/停止上报入参（v0.6）。"
      },
      "SkillOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "trigger_conditions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Trigger Conditions"
          },
          "guardrails": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Guardrails"
          },
          "steps": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Steps"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "action_type": {
            "type": "string",
            "title": "Action Type",
            "default": "guided_steps"
          },
          "estimated_duration": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Duration"
          },
          "completion_schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completion Schema"
          },
          "safety_constraints": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Safety Constraints"
          },
          "signed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signed By"
          },
          "signed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signed At"
          },
          "policy_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Version"
          },
          "review_evidence": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Evidence"
          },
          "revision": {
            "type": "integer",
            "title": "Revision",
            "default": 1
          },
          "supersedes_skill_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supersedes Skill Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "name",
          "version",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "SkillOut",
        "description": "技能包输出契约（PRD 契约点 4/5）。"
      },
      "SkillTransition": {
        "properties": {
          "new_status": {
            "type": "string",
            "enum": [
              "reviewed",
              "signed",
              "retired"
            ],
            "title": "New Status"
          },
          "policy_version": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Version"
          },
          "review_evidence": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Evidence"
          }
        },
        "type": "object",
        "required": [
          "new_status"
        ],
        "title": "SkillTransition",
        "description": "Skill 治理状态机转换入参。new_status 仅允许 reviewed/signed/retired。\n\n转入 signed 时必须提供 policy_version 与 review_evidence（契约点 5 重签校验）。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}